Edit page and reset password page added, reset password path working
